Move helpers.rs to fixtures/mod.rs
